+++ CHANGELOG.md
@@ -96,6 +96,10 @@ See these [`1.1.x` to `1.2.x` upgrade instructions](https://gist.github.com/chri
   * [phoenix.new] Provide `local.phoenix` task
   * [phoenix.digest] Add `?vsn=d` to digested assets
 
+* Enhancements
+  * [Router] Router macros now default to an empty list for `plug_opts`. This
+    allows any module satifying the plug behaviour to be used as a controller.
+
 ## 1.1.3 (2016-1-20)
 
 * Enhancements